Fixes State map not getting properly updated when copying StorageItems.
This is a workaround for a different bug with the State#entrySet / State.Entry#setValue API where calling setValue does not properly update the State and a subsequent call to State#save would lose any changes made. The workaround is to loop over the State#keySet and just call State#put directly.