Changes SqlQuery to skip the IGNORE INDEX (PRIMARY) hint when appropriate.
See: da721bd. If any predicate limits the resultset, it is more
efficient to meet those conditions before pulling in the typeId. If all
of the predicates are NOT IN or != or IS NOT NULL, the resultset is
still effectively unlimited, and the PRIMARY key is the optimal
execution plan.